# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Post, cancel, filter and retrieve trade(s) based on request parameters.",
    "title": "Trade",
    "version": "2.0.0",
    "x-aladdin-api-id": "agraph.trading.trade_processing.trade.TradeAPI",
    "x-aladdin-api-stability": "alpha",
    "x-aladdin-spec-id": "agraph.trading.trade_processing.trade.v2.TradeAPI"
  },
  "paths": {
    "/longrunningoperations/{id}": {
      "get": {
        "description": "Poll for the result of a long-running operation.",
        "operationId": "TradeAPI_GetLongrunningOperation",
        "parameters": [
          {
            "description": "A unique identifier for this request.",
            "in": "header",
            "name": "VND.com.blackrock.Request-ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          {
            "description": "Timestamp assigned to this request at origin, in \"HTTP-date\" format as defined by RFC 7231.",
            "in": "header",
            "name": "VND.com.blackrock.Origin-Timestamp",
            "required": true,
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "description": "Id of the LongrunningOperation",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "headers": {
              "VND.com.blackrock.Request-ID": {
                "description": "The unique identifier assigned by the client to the request that originated this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-ID": {
                "description": "A Unique Identifier for this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-Timestamp": {
                "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
                "example": "2026-07-01 12:26:13.499",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "x-aladdin-lro": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1LongrunningOperation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          }
        },
        "summary": "Get the long-running operation results",
        "x-access-type": "READ",
        "x-quota-limit": 1000
      }
    },
    "/trades:batchCancel": {
      "post": {
        "description": "Cancels multiple trades based on a list of TradeKeys and a shared CancelTradeConfig, returning a long‑running operation that exposes per‑trade cancel results when completed.",
        "operationId": "TradeAPI_BatchCancelTrades",
        "parameters": [
          {
            "description": "A unique identifier for this request.",
            "in": "header",
            "name": "VND.com.blackrock.Request-ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          {
            "description": "Timestamp assigned to this request at origin, in \"HTTP-date\" format as defined by RFC 7231.",
            "in": "header",
            "name": "VND.com.blackrock.Origin-Timestamp",
            "required": true,
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tradev2BatchCancelTradesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "headers": {
              "VND.com.blackrock.Request-ID": {
                "description": "The unique identifier assigned by the client to the request that originated this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-ID": {
                "description": "A Unique Identifier for this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-Timestamp": {
                "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
                "example": "2026-07-01 12:26:13.499",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "x-aladdin-lro": {
              "meta-data-type": {
                "description": "Meta data related to a BatchCancelTradesResponse.",
                "type": "object"
              },
              "response-type": {
                "description": "Returns the per-trade outcomes for a batch cancel operation, with each TradeResult indicating whether a given trade was successfully cancelled or failed.",
                "properties": {
                  "results": {
                    "description": "Contains the results of the batch cancel trades operation.",
                    "items": {
                      "$ref": "#/components/schemas/tradev2TradeResult"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1LongrunningOperation_BatchCancelTrades"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          }
        },
        "summary": "Cancel trades",
        "x-access-type": "WRITE",
        "x-quota-limit": 250
      }
    },
    "/trades:batchPost": {
      "post": {
        "description": "Post or update multiple trades, returning a long‑running operation that exposes per‑trade post or update results when completed.",
        "operationId": "TradeAPI_BatchPostTrades",
        "parameters": [
          {
            "description": "A unique identifier for this request.",
            "in": "header",
            "name": "VND.com.blackrock.Request-ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          {
            "description": "Timestamp assigned to this request at origin, in \"HTTP-date\" format as defined by RFC 7231.",
            "in": "header",
            "name": "VND.com.blackrock.Origin-Timestamp",
            "required": true,
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tradev2BatchPostTradesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "headers": {
              "VND.com.blackrock.Request-ID": {
                "description": "The unique identifier assigned by the client to the request that originated this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-ID": {
                "description": "A Unique Identifier for this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-Timestamp": {
                "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
                "example": "2026-07-01 12:26:13.499",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "x-aladdin-lro": {
              "meta-data-type": {
                "description": "Meta data related to a BatchPostTradesResponse.",
                "type": "object"
              },
              "response-type": {
                "description": "Represents the response for the BatchPostTrades, indicating posting status for each trade.",
                "properties": {
                  "results": {
                    "description": "Contains the results of the batch post trades operation.",
                    "items": {
                      "$ref": "#/components/schemas/tradev2TradeResult"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1LongrunningOperation_BatchPostTrades"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          }
        },
        "summary": "Post or update trades",
        "x-access-type": "WRITE",
        "x-quota-limit": 250
      }
    },
    "/trades:cancel": {
      "post": {
        "description": "Cancels a single trade identified by its Trade Key (Invnum and Portfolio Reference).",
        "operationId": "TradeAPI_CancelTrade",
        "parameters": [
          {
            "description": "A unique identifier for this request.",
            "in": "header",
            "name": "VND.com.blackrock.Request-ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          {
            "description": "Timestamp assigned to this request at origin, in \"HTTP-date\" format as defined by RFC 7231.",
            "in": "header",
            "name": "VND.com.blackrock.Origin-Timestamp",
            "required": true,
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tradev2CancelTradeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "headers": {
              "VND.com.blackrock.Request-ID": {
                "description": "The unique identifier assigned by the client to the request that originated this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-ID": {
                "description": "A Unique Identifier for this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-Timestamp": {
                "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
                "example": "2026-07-01 12:26:13.499",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tradev2CancelTradeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          }
        },
        "summary": "Cancel trade",
        "x-access-type": "WRITE",
        "x-quota-limit": 250
      }
    },
    "/trades:filter": {
      "post": {
        "description": "Filter trades using flexible portfolio, date/time, trade, asset, and broker filters, returns a paginated result set that can be enriched with related trade data via options and expands.",
        "operationId": "TradeAPI_FilterTrades",
        "parameters": [
          {
            "description": "A unique identifier for this request.",
            "in": "header",
            "name": "VND.com.blackrock.Request-ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          {
            "description": "Timestamp assigned to this request at origin, in \"HTTP-date\" format as defined by RFC 7231.",
            "in": "header",
            "name": "VND.com.blackrock.Origin-Timestamp",
            "required": true,
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tradev2FilterTradesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "headers": {
              "VND.com.blackrock.Request-ID": {
                "description": "The unique identifier assigned by the client to the request that originated this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-ID": {
                "description": "A Unique Identifier for this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-Timestamp": {
                "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
                "example": "2026-07-01 12:26:13.499",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tradev2FilterTradesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          }
        },
        "summary": "Filter trade by criteria",
        "x-access-type": "READ",
        "x-quota-limit": 1000
      }
    },
    "/trades:post": {
      "post": {
        "description": "Creates a new trade or updates an existing one in Aladdin using the supplied payload. Returns a snippet of the updated trade details on success, or an error object describing the failure.",
        "operationId": "TradeAPI_PostTrade",
        "parameters": [
          {
            "description": "A unique identifier for this request.",
            "in": "header",
            "name": "VND.com.blackrock.Request-ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          {
            "description": "Timestamp assigned to this request at origin, in \"HTTP-date\" format as defined by RFC 7231.",
            "in": "header",
            "name": "VND.com.blackrock.Origin-Timestamp",
            "required": true,
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tradev2PostTradeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "headers": {
              "VND.com.blackrock.Request-ID": {
                "description": "The unique identifier assigned by the client to the request that originated this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-ID": {
                "description": "A Unique Identifier for this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-Timestamp": {
                "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
                "example": "2026-07-01 12:26:13.499",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tradev2PostTradeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          }
        },
        "summary": "Post or update trade",
        "x-access-type": "WRITE",
        "x-quota-limit": 250
      }
    },
    "/trades:retrieve": {
      "post": {
        "description": "Retrieves trades by trade keys, external references, or order identifiers (optionally with expanded asset data) and an overall processing status.",
        "operationId": "TradeAPI_RetrieveTrades",
        "parameters": [
          {
            "description": "A unique identifier for this request.",
            "in": "header",
            "name": "VND.com.blackrock.Request-ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          {
            "description": "Timestamp assigned to this request at origin, in \"HTTP-date\" format as defined by RFC 7231.",
            "in": "header",
            "name": "VND.com.blackrock.Origin-Timestamp",
            "required": true,
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tradev2RetrieveTradesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "headers": {
              "VND.com.blackrock.Request-ID": {
                "description": "The unique identifier assigned by the client to the request that originated this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-ID": {
                "description": "A Unique Identifier for this response.",
                "schema": {
                  "type": "string",
                  "format": "UUID"
                }
              },
              "VND.com.blackrock.Response-Timestamp": {
                "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
                "example": "2026-07-01 12:26:13.499",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tradev2RetrieveTradesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          }
        },
        "summary": "Retrieve trade by key criteria",
        "x-access-type": "READ",
        "x-quota-limit": 1000
      }
    }
  },
...
  "servers": [
    {
      "url": "https://tcw.blackrock.com/api/trading/trade-processing/trade/v2"
    }
  ],
  "components": {
    "responses": {
      "400": {
        "description": "Aladdin® will not process the request due to an apparent client error.",
        "headers": {
          "VND.com.blackrock.Request-ID": {
            "description": "The unique identifier assigned by the client to the request that originated this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-ID": {
            "description": "A Unique Identifier for this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-Timestamp": {
            "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "description": "Defines an error that occurred.",
              "properties": {
                "code": {
                  "description": "A short mnemonic reference code for the error.",
                  "maxLength": 40,
                  "type": "string"
                },
                "message": {
                  "description": "A human readable description of the error.",
                  "type": "string"
                }
              },
              "required": [
                "code",
                "message"
              ],
              "type": "object"
            }
          }
        }
      },
      "401": {
        "description": "Aladdin® will not process the request as the necessary credentials were not provided or were invalid.",
        "headers": {
          "VND.com.blackrock.Request-ID": {
            "description": "The unique identifier assigned by the client to the request that originated this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-ID": {
            "description": "A Unique Identifier for this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-Timestamp": {
            "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "description": "Defines an error that occurred.",
              "properties": {
                "code": {
                  "description": "A short mnemonic reference code for the error.",
                  "maxLength": 40,
                  "type": "string"
                },
                "message": {
                  "description": "A human readable description of the error.",
                  "type": "string"
                }
              },
              "required": [
                "code",
                "message"
              ],
              "type": "object"
            }
          }
        }
      },
      "403": {
        "description": "Aladdin® understood the request but is refusing to authorize it.",
        "headers": {
          "VND.com.blackrock.Request-ID": {
            "description": "The unique identifier assigned by the client to the request that originated this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-ID": {
            "description": "A Unique Identifier for this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-Timestamp": {
            "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "description": "Defines an error that occurred.",
              "properties": {
                "code": {
                  "description": "A short mnemonic reference code for the error.",
                  "maxLength": 40,
                  "type": "string"
                },
                "message": {
                  "description": "A human readable description of the error.",
                  "type": "string"
                }
              },
              "required": [
                "code",
                "message"
              ],
              "type": "object"
            }
          }
        }
      },
      "429": {
        "description": "Aladdin® is refusing to process the request because the client has exceeded the rate limit.",
        "headers": {
          "VND.com.blackrock.Request-ID": {
            "description": "The unique identifier assigned by the client to the request that originated this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-ID": {
            "description": "A Unique Identifier for this response.",
            "schema": {
              "type": "string",
              "format": "UUID"
            }
          },
          "VND.com.blackrock.Response-Timestamp": {
            "description": "Timestamp assigned to this response, in \"HTTP-date\" format as defined by RFC 7231.",
            "example": "2026-07-01 12:26:13.499",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "description": "Defines an error that occurred.",
              "properties": {
                "code": {
                  "description": "A short mnemonic reference code for the error.",
                  "maxLength": 40,
                  "type": "string"
                },
                "message": {
                  "description": "A human readable description of the error.",
                  "type": "string"
                }
              },
              "required": [
                "code",
                "message"
              ],
              "type": "object"
            }
          }
        }
      }
    },
    "schemas": {
...
      "tradev2BatchCancelTradesRequest": {
        "description": "Specifies a batch of trades to be cancelled, along with configuration controlling how the cancellations are processed.",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/v2CancelTradeConfig"
          },
          "tradeKeys": {
            "description": "List of trades to be cancelled. At least one trade must be specified.\nUsers are advised to use small batch sizes and increase parallelism for performance.",
            "items": {
              "$ref": "#/components/schemas/v2TradeKey"
            },
            "required": [
              "trade_keys"
            ],
            "type": "array"
          }
        },
        "required": [
          "tradeKeys"
        ],
        "type": "object"
      },
...
      "v1LongrunningOperation": {
        "description": "LongrunningOperation represents a long-running operation that is the result\nof a network API call.",
        "properties": {
          "done": {
            "description": "If the value is `false`, it means the operation is still in progress.\nIf true, the operation is completed, and either `error` or `response` is\navailable.",
            "type": "boolean"
          },
          "error": {
            "$ref": "#/components/schemas/googlerpcStatus"
          },
          "id": {
            "description": "The server-assigned id, which is a UUID of 36 characters in length,\nonly unique within the same service that originally returns it.",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/Any"
          },
          "response": {
            "$ref": "#/components/schemas/Any"
          }
        },
        "type": "object"
      },
...
    }
  }
}